[Bug  Book accommodation
Initially, the getBookings routes were splitted into two which are
getUserBookings and getAllBookings. This may pose issue on frontend
design as regards which route should be used.

This commit makes sure
that all users can access the same route to get bookings and are served
data based on their role.